requirements.txt
Flask>=3.1.1
SQLAlchemy>=2.0.41
Flask-SQLAlchemy>=3.1.1
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
requests>=2.32.3